Define the color scale for "custom_data"
Defines the color scale for the "custom_data" coloring such that the
colors always match the values. Without this hardcoded scale, it is
possible for the legend to swap the values of "yes" and "no" and their
corresponding colors. This could cause a mismatch between the
screenshots in our documentation and what the user sees.